// apps/github/src/achievements/lib/achievement-type.ts
export enum AchievementType {
  PALINDROME = 1,
  FIRST_PR = 2,
  FIRST_ISSUE = 3,
  PR_MORE_DELETIONS = 4,
  PR_ONLY_DELETIONS = 5,
  FIRST_PROJECT_PR = 7,
  ONE_LINE_CHANGE = 8,
  QUICK_PR_APPROVAL = 9,
  RECEIVED_QUICK_PR_APPROVAL = 10,
  PR_MERGE_NO_APPROVAL = 11,
  MILESTONE_PERSONAL_PR_MERGED_5 = 12,
  MILESTONE_PERSONAL_PR_MERGED_10 = 13,
  MILESTONE_PERSONAL_PR_MERGED_25 = 14,
  MILESTONE_PERSONAL_PR_MERGED_50 = 15,
  MILESTONE_PERSONAL_PR_MERGED_100 = 16,
  MILESTONE_PERSONAL_PR_MERGED_200 = 17,
  MILESTONE_REPO_PR_MERGED_10 = 18,
  MILESTONE_REPO_PR_MERGED_25 = 19,
  MILESTONE_REPO_PR_MERGED_50 = 20,
  MILESTONE_REPO_PR_MERGED_100 = 21,
  MILESTONE_REPO_PR_MERGED_200 = 22,
  MILESTONE_REPO_PR_MERGED_300 = 23,
  PR_MULTIPLE_CONTRIBUTORS = 24,
  PR_MERGE_EQUAL_LINES_ADDED_DELETED = 28,
  PR_MERGE_BY_NOT_AUTHOR = 29,
  MILESTONE_PERSONAL_PR_REVIEWED_1 = 30,
  MILESTONE_PERSONAL_PR_REVIEWED_10 = 31,
  MILESTONE_PERSONAL_PR_REVIEWED_25 = 32,
  MILESTONE_PERSONAL_PR_REVIEWED_50 = 33,
  MILESTONE_PERSONAL_PR_REVIEWED_100 = 34,
  MILESTONE_PERSONAL_PR_REVIEWED_200 = 35,
}